% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/condition.R
\name{condition}
\alias{condition}
\title{Produces conditional probability from a joint one}
\usage{
condition(p, cond)
}
\arguments{
\item{p}{A joint probability table}

\item{cond}{A vector providing the subscripts which the probability will be conditioned over.  For instance, 1 indicates the first vector, 2 indicates the second vector, ....}
}
\value{
A conditional probability table
}
\description{
\code{condition} computes the conditional probability from a joint probability talbe.
}
\details{
The probability table must be represent a joint probability table which sums to 1.
}
\examples{
x = c(1,2,3,3)
y = c(1,1,2,2)
p = prob(x,y) # P(x,y)
q = condition(p,1) # P(y|x)
}
\seealso{
\code{\link{prob}} and \code{\link{marginalize}}
}

