% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_entropy.R
\name{effective_entropy}
\alias{effective_entropy}
\title{Compute effective entropy}
\usage{
effective_entropy(D, f, Nloop = 4000, Nfine = 300, pa = -4, pb = 3)
}
\arguments{
\item{D}{A dissimilarity matrix}

\item{f}{Weigh array}

\item{Nloop}{Number of loops}

\item{Nfine}{finesse of powers}

\item{pa}{initial power}

\item{pb}{final power}
}
\description{
Compute effective entropy
}
