\name{entropy-internal}

\alias{nsbsave}
\alias{nsbload}

\alias{get.lambda.shrink}

\alias{pvt.chisq.pval}


\title{Internal entropy functions}
\description{
  Internal entropy functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
