% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Calculates joint entropy from joint probability table}
\usage{
entropy(p)
}
\arguments{
\item{p}{A probability table}
}
\description{
This function computes entropy (in bits) from a probability table.
}
\details{
The probability table p must sum to 1.
}
\examples{
x = c(0,1,2,0)
p = prob(x) # P(x)
h = entropy(p) # H(x)
}
\seealso{
\code{\link{prob}}
}

