% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/h.R
\name{h}
\alias{h}
\title{Calculates joint entropy from feature vectors}
\usage{
h(...)
}
\arguments{
\item{...}{A number of vectors which are same length.}
}
\description{
This function calculates entropy in bits from data vectors.
}
\examples{
N = 100
x = sample(1:2,size=N,replace=T)
y = sample(1:3,size=N,replace=T)
z = sample(1:4,size=N,replace=T)
h = h(x,y,z) # H(x,y,z)
}

