% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mi.R
\name{mi}
\alias{mi}
\title{Calculates mutual information from feature vectors}
\usage{
mi(x, y)
}
\arguments{
\item{x}{A data vector}

\item{y}{A data vector which is the same length as x}
}
\description{
This function calculates mutual information from two feature vectors.
}
\examples{
x = c(0,1,1,1)
y = c(1,0,1,0)
I = mi(x,y) # I>0
}
\seealso{
\code{\link{entropy}}
}

