% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{soft_clustering}
\alias{soft_clustering}
\title{Use soft clustering algorithm}
\usage{
soft_clustering(fr, Zr, Sr, nloops)
}
\arguments{
\item{fr}{Weight vector}

\item{Zr}{The Z matrix}

\item{Sr}{the S matrix}

\item{nloops}{the number of loops to execute}
}
\value{
a list containing rhos and Zs
}
\description{
Use soft clustering algorithm
}
