% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_entropy.R
\name{effective_entropy}
\alias{effective_entropy}
\title{Compute effective entropy}
\usage{
effective_entropy(
  dissimilarity,
  f,
  Nloop = 4000,
  Nfine = 300,
  pa = -4,
  pb = 3,
  power_crit = -0.115148
)
}
\arguments{
\item{dissimilarity}{A dissimilarity matrix}

\item{f}{Weigh array}

\item{Nloop}{Number of loops}

\item{Nfine}{finesse of powers}

\item{pa}{initial power}

\item{pb}{final power}

\item{power_crit}{change initial conditions}
}
\description{
Compute effective entropy
}
\examples{
M <- matrix(seq(9), nrow = 3)
f <- rowSums(M) / sum(M)
D <- dist(f*M)
effective_entropy(D,f)
}
